pyproject.toml
[project]
# 📦 Project metadata
name = "fastpi-project"
version = "0.1.0"
requires-python = ">=3.12"
# 🔗 Core runtime dependencies (minimal required for app)
dependencies = []

[tool.black]
line-length = 110

[tool.flake8]
max-line-length = 110
exclude = [".venv"]

[dependency-groups]
api = [
  "fastapi>=0.116.1",
  "python-multipart>=0.0.20",
  "pyyaml>=6.0.2",
]

db = [
  "sqlmodel>=0.0.20",
  "psycopg2-binary>=2.9.9",
  "python-dotenv>=1.0.1",
]

test = [
  "black",
  "flake8",
  "flake8-pyproject",
  "autopep8",
  "pytest",
  "uvicorn",
  "schemathesis==4.0.3",
  "prance==25.4.8.0",
  "openapi-spec-validator==0.7.2",
  "pytest-cov>=6.2.1",
  "isort",
  "mypy"
]